pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piply"
version = "0.1.0"
description = "Lightweight YAML-driven workflow orchestration framework"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Piply Team" }]
keywords = ["workflow", "orchestration", "pipeline", "yaml", "prefect", "dbt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "prefect>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "schedule>=1.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
]

[project.scripts]
piply = "piply.cli.main:app"

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "black>=23.0.0", "ruff>=0.0.290"]

[tool.setuptools.packages.find]
where = ["."]
include = ["piply*", "tests*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = []

[tool.ruff.per-file-ignores]
"tests/*" = ["E501"]
